Task 0

+++ python_async_function/0-basic_async_syntax.py
@@ -6,12 +6,12 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     '''
     This is a asynchronous coroutine with random delay between 0 and max_delay
 
     Parameters:
-    - max_delay (float)
+    - max_delay (int)
 
     Returns:
     - float: random delay between 0 and max_delay
